Restructure by context of things
