+++ crates/bevy_input/src/mouse/button.rs
@@ -1,3 +1,5 @@
+use crate::ElementState;
+
 /// A button on a mouse device
 #[derive(Debug, Hash, PartialEq, Eq, Clone, Copy)]
 #[cfg_attr(feature = "serialize", derive(serde::Serialize, serde::Deserialize))]
@@ -6,4 +8,11 @@ pub enum MouseButton {
     Right,
     Middle,
     Other(u16),
+}
+
+/// A mouse button input event
+#[derive(Debug, Clone)]
+pub struct MouseButtonInput {
+    pub button: MouseButton,
+    pub state: ElementState,
 }
