+++ crates/bevy_input/src/mouse/mod.rs
@@ -1,6 +1,6 @@
 pub mod button;
-pub mod event;
+pub mod motion;
 pub mod scrolling;
 pub mod system;
 
-pub use crate::mouse::{button::*, event::*, scrolling::*, system::*};
+pub use crate::mouse::{button::*, motion::*, scrolling::*, system::*};
