+++ crates/bevy_input/src/mouse/motion.rs
@@ -1,0 +1,7 @@
+use bevy_math::Vec2;
+
+/// A mouse motion event
+#[derive(Debug, Clone)]
+pub struct MouseMotion {
+    pub delta: Vec2,
+}
