+++ crates/bevy_input/src/mouse/scrolling.rs
@@ -4,3 +4,12 @@ pub enum MouseScrollUnit {
     Line,
     Pixel,
 }
+
+/// A mouse scroll wheel event, where x represents horizontal scroll and y represents vertical
+/// scroll.
+#[derive(Debug, Clone)]
+pub struct MouseWheel {
+    pub unit: MouseScrollUnit,
+    pub x: f32,
+    pub y: f32,
+}
